fixup!: rename `index.d.ts` → `types.ts` per codemod limitation
