+++ recipes/correct-ts-specifiers/src/get-not-found-url.ts
@@ -1,6 +1,6 @@
 import { pathToFileURL } from 'node:url';
 
-import type { FSAbsolutePath, ResolvedSpecifier } from './index.d.ts';
+import type { FSAbsolutePath, ResolvedSpecifier } from './types.ts';
 
 export const getNotFoundUrl = (err: NodeJS.ErrnoException & { url?: FSAbsolutePath }) =>
 	pathToFileURL(err?.url ?? err.message.split("'")[1])?.href as ResolvedSpecifier;
